Setting up the DataFlags for active particles.
The idea here is that each ActiveParticle will accept a struct,
ActiveParticleFormationDataFlags, and then set the flags to true if it needs a
given type of supplemental data.  The active particle types will be iterated
over, the flags set, and then an instance of ActiveParticleFormationData will
be created.  All the flags that are set to 'true' indicate that a given piece
of supplemental data will have to be generated.  This will be done, and then
during a final pass the entire structure will be eliminated.

In pseudopythoncode, this will look like:

supp_data = {}
supp_data_flags = {}
for particle_type:
    particle_type.update_flags(supp_data_flags)
for flag in supp_data_flags:
    generate_data(flag, supp_data)
...generate particles...
for flag in supp_data_flags:
    destroy_data(flag, supp_data)

--HG--
branch : active_particles