This may end up being a failed experiment.
This uses a vector to pass new particles back to the ActiveParticleHandler,
where they are copied into a statically allocated list.  It is not yet clear to
me how the ->clear() call will interact with the underlying pointers.  The best
solution may end up being to allocate a MaximumNumberOfNewParticles pointer and
fill it progressively.

--HG--
branch : active_particles